deinline fmpz_vec IO
