+++ src/fmpz_vec/io.c
@@ -1,4 +1,5 @@
 /*
+    Copyright (C) 2008, 2009, 2010 William Hart
     Copyright (C) 2010 Sebastian Pancratz
 
     This file is part of FLINT.
@@ -9,10 +10,49 @@
     (at your option) any later version.  See <https://www.gnu.org/licenses/>.
 */
 
-#include "flint.h"
+#include <stdio.h>
 #include "gmpcompat.h"
 #include "fmpz.h"
 #include "fmpz_vec.h"
+
+/* printing *******************************************************************/
+
+/*
+    Recall the return value conventions for fputc (of type int)
+
+    ``If there are no errors, the same character that has been written is
+    returned.  If an error occurs, EOF is returned and the error indicator
+    is set''
+
+    where the EOF macro expands to a negative int, and flint_fprintf (of type int)
+
+    ``On success, the total number of characters written is returned.
+    On failure, a negative number is returned.''
+ */
+
+int _fmpz_vec_fprint(FILE * file, const fmpz * vec, slong len)
+{
+    int r;
+    slong i;
+
+    r = flint_fprintf(file, "%wd", len);
+    if ((len > 0) && (r > 0))
+    {
+        r = fputc(' ', file);
+        for (i = 0; (i < len) && (r > 0); i++)
+        {
+            r = fputc(' ', file);
+            if (r > 0)
+                r = fmpz_fprint(file, vec + i);
+        }
+    }
+
+    return r;
+}
+
+int _fmpz_vec_print(const fmpz * vec, slong len) { return _fmpz_vec_fprint(stdout, vec, len); }
+
+/* reading ********************************************************************/
 
 int _fmpz_vec_fread(FILE * file, fmpz ** vec, slong * len)
 {
@@ -69,3 +109,4 @@ int _fmpz_vec_fread(FILE * file, fmpz ** vec, slong * len)
     return 1;
 }
 
+int _fmpz_vec_read(fmpz ** vec, slong * len) { return _fmpz_vec_fread(stdin, vec, len); }
